BST done; still need update menus to new changes

+++ src/PayServiceController.cpp
@@ -118,7 +118,7 @@ unsigned long PayServiceController::getCreditCardNumber() {
 	return creditCardNumber;
 }
 
-int PayServiceController::getAmountToPay() {
+double PayServiceController::getAmountToPay() {
 	return amountToPay;
 	
 }
